+++ packages/components/Button/CHANGELOG.md
@@ -1,8 +1,25 @@
 # Change Log - @fluentui-react-native/button
 
-This log was last generated on Mon, 06 Jun 2022 19:20:10 GMT and should not be manually modified.
+This log was last generated on Wed, 15 Jun 2022 18:26:37 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.22.29
+
+Wed, 15 Jun 2022 18:26:37 GMT
+
+### Patches
+
+- Explicitly expose enableFocusRing (ruaraki@microsoft.com)
+
+## 0.22.28
+
+Mon, 13 Jun 2022 23:58:18 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.16.3
+- Bump @fluentui-react-native/pressable to v0.8.45
 
 ## 0.22.27