+++ packages/components/Callout/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/callout
 
-This log was last generated on Mon, 06 Jun 2022 19:20:10 GMT and should not be manually modified.
+This log was last generated on Thu, 16 Jun 2022 17:09:27 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.20.8
+
+Thu, 16 Jun 2022 17:09:27 GMT
+
+### Patches
+
+- apple: remove unneccesary build settings from podspec (sanajmi@microsoft.com)
+
+## 0.20.7
+
+Mon, 13 Jun 2022 23:58:18 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/interactive-hooks to v0.16.3
 
 ## 0.20.6